requirements.txt
pyyaml
nltk
spacy

bitarray
mmh3